lab1/backend/requirements.txt
djangorestframework
django-cors-headers
django-filter
psycopg[binary,pool]
markdown
faker
gunicorn